+++ src/OF_WAR.cpp
@@ -290,11 +290,12 @@ void FirmWar::disp_build_menu(int refreshFlag)
 		if( !added_count || queue_weapon_selected >= added_count)
 			queue_weapon_selected = -1;
 
-		button_cancel.paint(x, y, x+BUILD_BUTTON_WIDTH-1, y+BUILD_BUTTON_HEIGHT*3/4,
+		button_cancel.paint(x, y, x+Ambition::Draw::calculateDoneButtonWidth(BUILD_BUTTON_WIDTH, this)-1, y+BUILD_BUTTON_HEIGHT*3/4,
 		ButtonCustom::disp_text_button_func, ButtonCustomPara((void*)_("Done"),0) );
 	}
 	// ###### end Gilbert 10/9 ########//
 
+	Ambition::Building::printProductionQueueTotals(this, refreshFlag, button_cancel.x2, button_cancel.y1, button_cancel.y2);
 }
 //----------- End of function FirmWar::disp_build_menu -----------//
 
@@ -650,6 +651,8 @@ static void i_disp_build_button(ButtonCustom *button, int repaintBody)
 			FirmWar *warFactory = (FirmWar *)queueButton->custom_para.ptr;
 			int techLevel = unitInfo->get_nation_tech_level(warFactory->nation_recno);
 
+			Ambition::Draw::warMachineTechnologyLevel(&techLevel, x1 + BUILD_UNIT_ICON_OFFSET_X, y1 + BUILD_UNIT_ICON_OFFSET_Y, unitInfo->get_large_icon_ptr(0));
+
 			if( techLevel > 1 )
 			{
 				str += " ";
@@ -661,6 +664,8 @@ static void i_disp_build_button(ButtonCustom *button, int repaintBody)
 			str += "<";
 
 		font_bible.put( x1+BUILD_UNIT_NAME_OFFSET_X, y1+BUILD_UNIT_NAME_OFFSET_Y, str );
+
+		Ambition::Draw::buttonCost(font_bible, unitInfo->build_cost, x1, y1 + BUILD_UNIT_NAME_OFFSET_Y, x2 - 48, y2, -1);
 	}
 
 	// display small button